Fix the 0ad dependency to its latest format.
This issue follows https://github.com/0ADMods/no-blood-and-gore-mod/issues/3 , bringing the mod to be compatible and publishable to the official 0ad directory.

The reason why this change is needed is that 0ad changed its naming policy, and seemingly updated its version, too. It's now specified as "0.27.0" as opposed to the old "0.0.27".